Add Notes to Change History
