+++ src/dotnet/APIView/APIViewWeb/LeanModels/ChangeHistory.cs
@@ -12,11 +12,49 @@ namespace APIViewWeb.LeanModels
         Modified
     }
 
+    [JsonConverter(typeof(StringEnumConverter))]
+    public enum ReviewChangeAction
+    {
+        Subscribed,
+        UnSubScribed,
+        ApprovedForFirstRelease,
+        RevertedFirstReleaseApproval,
+        Closed,
+        ReOpened,
+        Deleted,
+        Undeleted
+    }
+
+    [JsonConverter(typeof(StringEnumConverter))]
+    public enum RevisionChangeAction
+    {
+        Approved = 0,
+        RevertedApproval,
+        Deleted,
+        Undeleted
+    }
+
     public class AICommentChangeHistoryModel
     {
-        [JsonConverter(typeof(StringEnumConverter))]
         public AICommentChangeAction ChangeAction { get; set; }
         public string User { get; set; }
         public DateTime ChangeDateTime { get; set; }
+    }
+
+    public class ReviewChangeHistoryModel
+    {
+        public ReviewChangeAction ChangeAction { get; set; }
+        public string User { get; set; }
+        public DateTime ChangeDateTime { get; set; }
+        public string Notes { get; set; }
+
+    }
+
+    public class RevisionChangeHistoryModel
+    {
+        public RevisionChangeAction ChangeAction { get; set; }
+        public string User { get; set; }
+        public DateTime ChangeDateTime { get; set; }
+        public string Notes { get; set; }
     }
 }
