+++ src/dotnet/APIView/APIViewWeb/Models/ReviewRevisionModel.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-using APIViewWeb.Models;
+using APIViewWeb.LeanModels;
 using Newtonsoft.Json;
 
 namespace APIViewWeb
